web/requirements.txt
pymongo
together
flask
gunicorn
flask_cors
python-dotenv
pypdf2
openai